Feature/permissions (#52)
* group endpoints added

* EoY wip

* wip

* task endpoints complete

* authentication complete and data migration complete

* removed ALL authorization from services and removed unused controllers

* remove unused code

* adding membership end points wip

* finish membership endpoints

* add user update endpoint to controller

* fix groupmembership endpoints

* add mappings

* fix get results by user

* added functionality for non-sysadmin users to get all of the scenarios and scenario templates that they have access to

* add permissions to each scenario and template

* add membership signalr

* Add missing document markings

* fix codeql-analysis action

* fix errors on first load of new user (#53)

---------

+++ Steamfitter.Api/Controllers/ScenarioTemplateController.cs
@@ -6,22 +6,24 @@ using System.Collections.Generic;
 using System.Net;
 using System.Threading;
 using STT = System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Steamfitter.Api.Infrastructure.Extensions;
+using Steamfitter.Api.Data;
+using Steamfitter.Api.Infrastructure.Authorization;
 using Steamfitter.Api.Infrastructure.Exceptions;
 using Steamfitter.Api.Services;
 using SAVM = Steamfitter.Api.ViewModels;
 using Swashbuckle.AspNetCore.Annotations;
+using Steamfitter.Api.ViewModels;
+using System.Linq;
 
 namespace Steamfitter.Api.Controllers
 {
     public class ScenarioTemplateController : BaseController
     {
         private readonly IScenarioTemplateService _scenarioTemplateService;
-        private readonly IAuthorizationService _authorizationService;
+        private readonly ISteamfitterAuthorizationService _authorizationService;
 
-        public ScenarioTemplateController(IScenarioTemplateService scenarioTemplateService, IAuthorizationService authorizationService)
+        public ScenarioTemplateController(IScenarioTemplateService scenarioTemplateService, ISteamfitterAuthorizationService authorizationService)
         {
             _scenarioTemplateService = scenarioTemplateService;
             _authorizationService = authorizationService;
@@ -41,45 +43,28 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "getScenarioTemplates")]
         public async STT.Task<IActionResult> Get(CancellationToken ct)
         {
-            var list = await _scenarioTemplateService.GetAsync(ct);
+            IEnumerable<SAVM.ScenarioTemplate> list;
+            // get ALL scenario templates
+            if (await _authorizationService.AuthorizeAsync([SystemPermission.ViewScenarioTemplates], ct))
+            {
+                list = await _scenarioTemplateService.GetAsync(ct);
+            }
+            // get scenario templates the user can access
+            else if (await _authorizationService.AuthorizeAsync([SystemPermission.CreateScenarioTemplates], ct))
+            {
+                list = await _scenarioTemplateService.GetMineAsync(ct);
+            }
+            // return forbidden exception
+            else
+            {
+                throw new ForbiddenException();
+            }
+
+            AddPermissions(list);
+
             return Ok(list);
+
         }
-
-        // /// <summary>
-        // /// Gets all ScenarioTemplates for a User
-        // /// </summary>
-        // /// <remarks>
-        // /// Returns all ScenarioTemplates where the specified User is a member of at least one of it's Teams
-        // /// <para />
-        // /// Accessible to a SuperUser or the specified User itself
-        // /// </remarks>
-        // /// <returns></returns>
-        // [HttpGet("users/{id}/scenarioTemplates")]
-        // [ProducesResponseType(typeof(IEnumerable<ScenarioTemplate>), (int)HttpStatusCode.OK)]
-        // [SwaggerOperation(OperationId = "getUserScenarioTemplates")]
-        // public async STT.Task<IActionResult> GetByUserId(Guid id, CancellationToken ct)
-        // {
-        //     var list = await _scenarioTemplateService.GetByUserIdAsync(id, ct);
-        //     return Ok(list);
-        // }
-
-        // /// <summary>
-        // /// Gets all ScenarioTemplates for the current User
-        // /// </summary>
-        // /// <remarks>
-        // /// Returns all ScenarioTemplates where the current User is a member of at least one of it's Teams
-        // /// <para />
-        // /// Accessible only to the current User.
-        // /// <para/>
-        // /// This is a convenience endpoint and simply returns a 302 redirect to the fully qualified users/{id}/scenarioTemplates endpoint
-        // /// </remarks>
-        // [HttpGet("me/scenarioTemplates")]
-        // [ProducesResponseType(typeof(IEnumerable<ScenarioTemplate>), (int)HttpStatusCode.OK)]
-        // [SwaggerOperation(OperationId = "getMyScenarioTemplates")]
-        // public async STT.Task<IActionResult> GetMy(CancellationToken ct)
-        // {
-        //     return RedirectToAction(nameof(this.GetByUserId), new { id = User.GetId() });
-        // }
 
         /// <summary>
         /// Gets a specific ScenarioTemplate by id
@@ -97,10 +82,12 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "getScenarioTemplate")]
         public async STT.Task<IActionResult> Get(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<SAVM.ScenarioTemplate>(id, [SystemPermission.ViewScenarioTemplates], [ScenarioTemplatePermission.ViewScenarioTemplate], ct))
+                throw new ForbiddenException();
+
             var scenarioTemplate = await _scenarioTemplateService.GetAsync(id, ct);
 
-            if (scenarioTemplate == null)
-                throw new EntityNotFoundException<SAVM.ScenarioTemplate>();
+            AddPermissions(scenarioTemplate);
 
             return Ok(scenarioTemplate);
         }
@@ -120,7 +107,12 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "createScenarioTemplate")]
         public async STT.Task<IActionResult> Create([FromBody] SAVM.ScenarioTemplateForm scenarioTemplateForm, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.CreateScenarioTemplates], ct))
+                throw new ForbiddenException();
+
             var createdScenarioTemplate = await _scenarioTemplateService.CreateAsync(scenarioTemplateForm, ct);
+            AddPermissions(createdScenarioTemplate);
+
             return CreatedAtAction(nameof(this.Get), new { id = createdScenarioTemplate.Id }, createdScenarioTemplate);
         }
 
@@ -139,15 +131,21 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "copyScenarioTemplate")]
         public async STT.Task<IActionResult> Copy(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync([SystemPermission.CreateScenarioTemplates], ct)
+                || !await _authorizationService.AuthorizeAsync<SAVM.ScenarioTemplate>(id, [SystemPermission.ViewScenarioTemplates], [ScenarioTemplatePermission.ViewScenarioTemplate], ct))
+                throw new ForbiddenException();
+
             var newScenarioTemplate = await _scenarioTemplateService.CopyAsync(id, ct);
+            AddPermissions(newScenarioTemplate);
+
             return CreatedAtAction(nameof(this.Get), new { id = newScenarioTemplate.Id }, newScenarioTemplate);
         }
 
         /// <summary>
-        /// Updates an ScenarioTemplate
+        /// Updates a ScenarioTemplate
         /// </summary>
         /// <remarks>
-        /// Updates an ScenarioTemplate with the attributes specified
+        /// Updates a ScenarioTemplate with the attributes specified
         /// <para />
         /// Accessible only to a SuperUser or a User on an Admin Team within the specified ScenarioTemplate
         /// </remarks>
@@ -159,15 +157,20 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "updateScenarioTemplate")]
         public async STT.Task<IActionResult> Update([FromRoute] Guid id, [FromBody] SAVM.ScenarioTemplateForm scenarioTemplateForm, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<SAVM.ScenarioTemplate>(id, [SystemPermission.EditScenarioTemplates], [ScenarioTemplatePermission.EditScenarioTemplate], ct))
+                throw new ForbiddenException();
+
             var updatedScenarioTemplate = await _scenarioTemplateService.UpdateAsync(id, scenarioTemplateForm, ct);
+            AddPermissions(updatedScenarioTemplate);
+
             return Ok(updatedScenarioTemplate);
         }
 
         /// <summary>
-        /// Deletes an ScenarioTemplate
+        /// Deletes a ScenarioTemplate
         /// </summary>
         /// <remarks>
-        /// Deletes an ScenarioTemplate with the specified id
+        /// Deletes a ScenarioTemplate with the specified id
         /// <para />
         /// Accessible only to a SuperUser or a User on an Admin Team within the specified ScenarioTemplate
         /// </remarks>
@@ -178,8 +181,26 @@ namespace Steamfitter.Api.Controllers
         [SwaggerOperation(OperationId = "deleteScenarioTemplate")]
         public async STT.Task<IActionResult> Delete(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<SAVM.ScenarioTemplate>(id, [SystemPermission.ManageScenarioTemplates], [ScenarioTemplatePermission.ManageScenarioTemplate], ct))
+                throw new ForbiddenException();
+
             await _scenarioTemplateService.DeleteAsync(id, ct);
             return NoContent();
+        }
+
+        private void AddPermissions(IEnumerable<SAVM.ScenarioTemplate> list)
+        {
+            foreach (var item in list)
+            {
+                AddPermissions(item);
+            }
+        }
+
+        private void AddPermissions(SAVM.ScenarioTemplate item)
+        {
+            item.ScenarioTemplatePermissions =
+            _authorizationService.GetScenarioTemplatePermissions(item.Id).Select((m) => m.ToString())
+            .Concat(_authorizationService.GetSystemPermissions().Select((m) => m.ToString()));
         }
 
     }
